05-monitoring/requirements.txt
# python==3.12.9
pandas
evidently==0.6.0
psycopg2-binary
pyarrow